Initial commit: MVP built from bits and pieces produced over the past year
I hacked on this and similar projects on and off for awhile.  This commit is after taking bits from those projects
and some original work and producing a functional site, though one with a long todo list.  I lose git history for
those prior odds and ends, but that should be fine.  This iteration also moves the dev environment to ubuntu
rather than windows.