first draft of run-based randomness, not yet tested.
devdiary: let this be a lesson for the future.  Hemming and hawwing about the best way to do something just leads to delays.  Finally got around to writing on the basic version after a few wasted days.  If I come up with improvements (and I will if they're necessary), they'll be easier to write and reason about because I've done the simple thing.  Also, taking breaks is fine but its better if they're intentional